+++ src/components/emission/new/DetailedGES.module.css
@@ -1,0 +1,11 @@
+.questions {
+  gap: 1rem;
+}
+
+.selector {
+  flex: 2;
+}
+
+.input {
+  flex: 1;
+}
