+++ src/components/emission/new/EmissionPostForm.tsx
@@ -1,12 +1,15 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 import { useMemo } from 'react'
 import { UseFormReturn } from 'react-hook-form'
 import { useTranslations } from 'next-intl'
-import { FormTextField } from '@/components/form/TextField'
+import styles from './EmissionPostForm.module.css'
 import { PostType } from '@prisma/client'
 import { CreateEmissionCommand } from '@/services/serverFunctions/emission.command'
 import { Accordion, AccordionDetails, AccordionSummary, MenuItem } from '@mui/material'
-import DetailedGESFields from './DetailedGESForm'
+import DetailedGESFields from './DetailedGESFields'
+import { FormTextField } from '@/components/form/TextField'
 import { FormSelect } from '@/components/form/Select'
+import classNames from 'classnames'
 
 interface DetailedGESFieldsProps {
   form: UseFormReturn<CreateEmissionCommand>
@@ -17,33 +20,18 @@ interface DetailedGESFieldsProps {
 
 const EmissionPostForm = ({ detailedGES, form, index }: DetailedGESFieldsProps) => {
   const t = useTranslations('emissions.create')
-  const tPostType = useTranslations('postType')
+  const tPostType = useTranslations('emissions.postType')
 
-  const header = form.watch(`posts.${index}.name`) || `post ${index + 1}`
+  const header = form.watch(`posts.${index}.name`) || `${t('post')} ${index + 1}`
   const totalCo2 = form.watch(`posts.${index}.totalCo2`) || 0
 
-  const posttTypes = useMemo(
-    () => Object.values(PostType).sort((a, b) => tPostType(a).localeCompare(tPostType(b))),
-    [t],
-  )
-
-  // const emissionValues = form.watch([
-  //   `ch4b.${index}`,
-  //   `ch4f.${index}`,
-  //   `co2b.${index}`,
-  //   `co2f.${index}`,
-  //   `n2o.${index}`,
-  //   `sf6.${index}`,
-  //   `hfc.${index}`,
-  //   `pfc.${index}`,
-  //   `otherGES.${index}`,
-  // ])
+  const postTypes = useMemo(() => Object.values(PostType).sort((a, b) => tPostType(a).localeCompare(tPostType(b))), [t])
 
   return (
-    <Accordion key={index}>
+    <Accordion>
       <AccordionSummary expandIcon={<>+</>}>{header}</AccordionSummary>
       <AccordionDetails className="flex-col">
-        <div className="flex" style={{ gap: '1rem' }}>
+        <div className={classNames(styles['accordion-details-header'], 'flex')}>
           <FormTextField
             data-testid={`new-emission-post-${index}-name`}
             control={form.control}
@@ -53,14 +41,14 @@ const EmissionPostForm = ({ detailedGES, form, index }: DetailedGESFieldsProps) 
             label={t('name')}
           />
           <FormSelect
-            data-testid={`post${index}-type`}
+            data-testid={`post-${index}-type`}
+            className={styles['accordion-details-type-selector']}
             control={form.control}
             translation={t}
             label={t('postType')}
             name={`posts.${index}.type`}
-            style={{ minWidth: '10rem' }}
           >
-            {posttTypes.map((postType) => (
+            {postTypes.map((postType) => (
               <MenuItem key={postType} value={postType}>
                 {tPostType(postType)}
               </MenuItem>
@@ -69,7 +57,7 @@ const EmissionPostForm = ({ detailedGES, form, index }: DetailedGESFieldsProps) 
         </div>
 
         {detailedGES ? (
-          <DetailedGESFields detailedGES={detailedGES} control={form.control} index={index} />
+          <DetailedGESFields form={form} index={index} />
         ) : (
           <FormTextField
             data-testid="new-emission-totalCo2"
