--- src/components/emission/new/styles.module.css
@@ -1,3 +1,0 @@
-.questions {
-  gap: 1rem;
-}
